split the run_executor management command
