+++ executor/app/src/compute_horde_executor/executor/tests/integration/test_main_loop.py
@@ -1,4 +1,3 @@
-import asyncio
 import base64
 import io
 import json
@@ -13,17 +12,18 @@ from unittest import mock
 from unittest.mock import patch
 
 import httpx
+import pytest
 from compute_horde.protocol_messages import V0InitialJobRequest, V0JobFailedRequest
 from compute_horde.transport import StubTransport
 from compute_horde_core.certificate import generate_certificate_at
 from pytest_httpx import HTTPXMock
 from requests_toolbelt.multipart import decoder
 
+from compute_horde_executor.executor.job_runner import JobRunner
 from compute_horde_executor.executor.management.commands.run_executor import (
     Command,
-    JobRunner,
-    MinerClient,
 )
+from compute_horde_executor.executor.miner_client import MinerClient
 
 payload = "".join(random.choice(string.ascii_uppercase + string.digits) for _ in range(32))
 
@@ -73,45 +73,21 @@ def get_file_from_request(request):
     return parsed_data
 
 
+@pytest.fixture(autouse=True)
+@patch("compute_horde_executor.executor.job_driver.JobDriver.run_cve_2022_0492_check_or_fail")
+@patch(
+    "compute_horde_executor.executor.job_driver.JobDriver.run_nvidia_toolkit_version_check_or_fail"
+)
+@patch("compute_horde_executor.executor.job_driver.JobDriver.run_security_checks_or_fail")
+def mock_gpu_checks(cve_check, nvidia_toolkit_check, security_check):
+    yield
+
+
 class CommandTested(Command):
     def __init__(self, messages, *args, **kwargs):
         transport = StubTransport("test", messages)
         self.MINER_CLIENT_CLASS = partial(MinerClient, transport=transport)
         super().__init__(*args, **kwargs)
-
-    async def run_nvidia_toolkit_version_check_or_fail(self):
-        is_toolkit_installed = None
-
-        try:
-            process = await asyncio.create_subprocess_exec(
-                "nvidia-container-toolkit",
-                "--version",
-                stdout=asyncio.subprocess.PIPE,
-                stderr=asyncio.subprocess.PIPE,
-            )
-        except OSError:
-            is_toolkit_installed = False
-
-        if is_toolkit_installed is None:
-            try:
-                stdout, stderr = await asyncio.wait_for(process.communicate(), 5)
-            except TimeoutError:
-                process.kill()
-                is_toolkit_installed = False
-
-        if is_toolkit_installed is None and process.returncode != 0:
-            is_toolkit_installed = False
-
-        if is_toolkit_installed is None:
-            is_toolkit_installed = True
-
-        if is_toolkit_installed:
-            return await super().run_nvidia_toolkit_version_check_or_fail()
-        else:
-            logger.warning(
-                "NVIDIA Container Toolkit not installed - skipping safe toolkit version check in tests"
-            )
-            return True
 
 
 def test_main_loop_basic():
